view_server/requirements.txt
click==6.7
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.0
Flask==0.12.2
Flask-SocketIO==2.8.6
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
paho-mqtt==1.2.3
python-engineio==1.5.4
python-socketio==1.7.5
six==1.10.0
Werkzeug==0.12.2
